statmodkit/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
matplotlib>=3.4.0
seaborn>=0.11.0
optuna>=2.10.0
mlxtend>=0.19.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
joblib>=1.1.0
dask>=2022.2.0
dask-ml>=1.9.0
python-Levenshtein>=0.12.0
PyYAML>=6.0 